+++ ChangeLog.md
@@ -20,6 +20,8 @@ v4.25.8
   - Remove unused method `Sisimai::Rhost->list` #396
   - Implement `Sisimai::Rhost::Cox` for parsing bounce mails returned from Cox:
     https://cox.com/. Thanks to @meir-w #398
+  - Callback feature for each email file in Maildir/ #399
+    - `Sisimai->make`, `Sisimai->dump` receive a hook method at `c___` argument
 
 v4.25.7
 --------------------------------------------------------------------------------